requirements.txt
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
cffi==1.16.0
chardet==5.2.0
click==8.1.7
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
dnspython==2.6.1
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.110.0
fonttools==4.51.0
h11==0.14.0
httptools==0.6.1
idna==3.6
jose==1.0.0
kiwisolver==1.4.5
matplotlib==3.8.4
motor==3.3.2
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.1
reportlab==4.1.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
